Add files via upload
This project aims to predict the Relative Humidity (RH) of a household using a dataset. The dataset contains observations of a 5 month period: 11 January 2016 to 27 May 2016. The dataset is a combination of measurements of humidity and temperature from households and weather data. The house has 9 residual zonesand Humidity and temperature of it has been measured every 10 minutes